src/main/java/com/example/springbootdatajpaquerydsl/service/ProductService.java
package com.example.springbootdatajpaquerydsl.service;

import com.example.springbootdatajpaquerydsl.domain.category.CategoryRepository;
import com.example.springbootdatajpaquerydsl.domain.product.Product;
import com.example.springbootdatajpaquerydsl.domain.product.ProductRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class ProductService {

    private final ProductRepository productRepository;
    private final CategoryRepository categoryRepository;

    public Product addProduct(Product product) {
        return productRepository.save(product);
    }

    public List<Product> getProducts() {
        return productRepository.findAll();
    }

    public Product getProductById(Long id) {
        return productRepository.findById(id).get();
    }

    public Product getProductByName(String name) {
        return productRepository.findByName(name);
    }

    public Product updateProduct(Product product) {
        return productRepository.save(product);
    }


    public boolean removeProduct(Long id) {
        if(productRepository.existsById(id)) {
            productRepository.deleteById(id);
            return true;
        }

        return false;
    }
}
